Fixed styling, sticky notes are now randomly left right. Fixed flex wrap. Created a new component folder, got some errors when I renamed it.
