--- src/Components/styledComponents/Media.jsx
@@ -1,8 +1,0 @@
-export const Media = {
-  mobile: "(max-width: 500px)",
-  tablet: "(min-width: 501px) and (max-width: 768px)",
-  desktop: "(min-width: 769px) and (max-width: 1200px)",
-  widescreen: "(min-width: 1201px)",
-  smallerSizes: "(max-width: 900px)",
-  biggerSizes: "(min-width: 900px)"
-}
