+++ src/component/styledComponents/Media.jsx
@@ -1,0 +1,6 @@
+export const Media = {
+  mobile: "(max-width: 500px)",
+  tablet: "(min-width: 501px) and (max-width: 768px)",
+  desktop: "(min-width: 769px) and (max-width: 1600px)",
+
+}
